png: Properly handle chunk value > 768 bytes.
The decoder.tmp array field is only 768 bytes (for unknown reasons).
The existing decoder code was guarding against overflowing the array.
Unfortunately, when we added support for tEXt/zTXt chunks, we did
not enforce such a check.

This commit adds a *bytes.Buffer that is dynamically instantiated
and re-used for parsing such chunk values. Theoretically, the tmp
array could be re-used for this - but it would be complicated and
error-prone to do so.