pluck: Initial take.
This utility examines all of the files in a directory and (assuming
they can be parsed as pngs) extracts chunks matching the keyword
regex into a sensible directory structure.

Optionally, it can execute 'hexdump -C' against the data extracted
from png chunks. The output of hexdump is saved to a file alongside
the chunk file(s) and info file(s).